mysite/requirements.txt
dj-database-url>=0.5.0
Django>=2.0.8
gunicorn>=19.9.0
psycopg2>=2.7.5
psycopg2-binary>=2.7.5
wheel>=0.31.1
whitenoise>=3.3.1
reportlab>=3.5.0
django-import-export>=1.0.1
django-countries>=5.3.1
django-multiselectfield>=0.1.8
django-betterforms>=1.2
django-widget-tweaks>=1.4.2
django-filter>=2.0.0
django-formtools>=2.1
pycountry>=18.5.26
django-upload-path>=0.1.1
django-extensions>=2.1.4
django-storages>=1.7.1
